requirements.txt
Flask~=3.0.2
PyYAML~=6.0.1
psutil~=5.9.8
pid~=3.0.4
waitress~=3.0.0
